+++ changelog.md
@@ -30,10 +30,12 @@
 
 - guidelines added for the game
 
-## version 2.0
+## version 1.0
 
 - Word and Words class added to simplify codebase
 - wordlist updated and all words have hints
 - Hint functionality added for words
 - different points for normal words (1) and custom words (2 pts)
 - codebase refactored
+- ask how many rounds players want to play (i.e. winning point)
+- prompt for providing hint when 2 chances are left